bugfix: sub-minute cron timers was randomly fired since Qt coarse timers
forcing QTimers to Qt::PreciseTimer was done *almost* everywhere but
when setting a sub-minute timer. this has random effects, and does not
apply if the interval is below 2 seconds (because in this case precise
timers was selected implictly)

the bug did not happen on minute triggers like 0 * * * * * nor to seconds
or 2-seconds triggers like */2 * * * * * but did apply for e.g.
*/5 * * * * * or 15,45 * * * * *