cron trigger: fixed a DST-related severe freeze
the scheduler was freezed (no taskinstance starts) when computing next
execution time of a cron trigger both allowing sunday and forcing a start
hour after the transition begin, e.g. (0 0 14 * * 7) did freeze qron all
the week before a sunday when the time should fall back from 1 hour
between 02:00 a.m. CEST and 02:00 a.m. CET

the bug is very likely related to Qt 6.7 change in QDateTime::setTime()
which introduce TransitionResolution which so called LegacyBehavior is
not truly the same than the legacy behavior, and the bug appeared in
qron 1.16.4 when we switched from Qt 6.5 to Qt 6.8
https://doc.qt.io/qt-6/qdatetime.html#setTime
https://doc.qt.io/qt-6/qdatetime.html#TransitionResolution-enum